+++ ios/Classes/JPushPlugin.m
@@ -23,7 +23,7 @@
 
 
 #if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
-@interface JPushPlugin ()<JPUSHRegisterDelegate>
+@interface JPushPlugin ()<JPUSHRegisterDelegate,JPUSHInAppMessageDelegate>
 //在前台时是否展示通知
 @property(assign, nonatomic) BOOL unShow;
 @end
@@ -169,6 +169,10 @@ static NSMutableArray<FlutterResult>* getRidResults;
         result(@YES);
     } else if ([@"setAuth" isEqualToString:call.method]) {
         [self setAuth:call result:result];
+    } else if ([@"pageEnterTo" isEqualToString:call.method]) {
+        [self pageEnterTo:call result:result];
+    } else if ([@"pageLeave" isEqualToString:call.method]) {
+        [self pageLeave:call result:result];
     } else{
         result(FlutterMethodNotImplemented);
     }
@@ -186,6 +190,8 @@ static NSMutableArray<FlutterResult>* getRidResults;
     } else {
         [JPUSHService setLogOFF];
     }
+    
+    [JPUSHService setInAppMessageDelegate:self];
     
     [JPUSHService setupWithOption:_completeLaunchNotification
                            appKey:arguments[@"appKey"]
@@ -514,6 +520,19 @@ static NSMutableArray<FlutterResult>* getRidResults;
     result(@YES);
 }
 
+- (void)pageEnterTo:(FlutterMethodCall*)call result:(FlutterResult)result {
+    JPLog(@"pageEnterTo:%@",call.arguments);
+    NSString *pageName = call.arguments;
+    [JPUSHService pageEnterTo:pageName];
+    result(@YES);
+}
+
+- (void)pageLeave:(FlutterMethodCall*)call result:(FlutterResult)result {
+    JPLog(@"pageLeave:%@",call.arguments);
+    NSString *pageName = call.arguments;
+    [JPUSHService pageLeave:pageName];
+    result(@YES);
+}
 
 
 - (void)dealloc {
@@ -739,5 +758,28 @@ static NSMutableArray<FlutterResult>* getRidResults;
     formatDic[@"extras"] = extras;
     return formatDic;
 }
+
+
+#pragma mark - 应用内消息回调
+- (void)jPushInAppMessageDidShow:(JPushInAppMessage *)inAppMessage {
+    [_channel invokeMethod:@"onInAppMessageShow" arguments: [self convertInappMsg:inAppMessage]];
+}
+
+- (void)jPushInAppMessageDidClick:(JPushInAppMessage *)inAppMessage {
+    [_channel invokeMethod:@"onInAppMessageClick" arguments: [self convertInappMsg:inAppMessage]];
+}
+
+- (NSDictionary *)convertInappMsg:(JPushInAppMessage *)inAppMessage {
+    NSDictionary *result = @{
+        @"mesageId": inAppMessage.mesageId ?: @"",    // 消息id
+        @"title": inAppMessage.title ?:@"",       // 标题
+        @"content": inAppMessage.content ?: @"",    // 内容
+        @"target": inAppMessage.target ?: @[],      // 目标页面
+        @"clickAction": inAppMessage.clickAction ?: @"", // 跳转地址
+        @"extras": inAppMessage.extras ?: @{} // 附加字段
+    };
+    return result;
+}
+
 
 @end